+++ src/common/mac.c
@@ -193,12 +193,12 @@ _mac_init (mac_ctx *x, munge_mac_t md, const void *key, int keylen)
     }
     if ((e = gcry_md_open (&x->ctx, algo, GCRY_MD_FLAG_HMAC)) != 0) {
         log_msg (LOG_DEBUG, "gcry_md_open failed for MAC=%d HMAC: %s",
-            md, gcry_strerror (e));
+            (int) md, gcry_strerror (e));
         return -1;
     }
     if ((e = gcry_md_setkey (x->ctx, key, keylen)) != 0) {
         log_msg (LOG_DEBUG, "gcry_md_setkey failed for MAC=%d HMAC: %s",
-            md, gcry_strerror (e));
+            (int) md, gcry_strerror (e));
         return -1;
     }
     /*  Bypass mac_size() since md->algo mapping has already been computed.
@@ -262,12 +262,12 @@ _mac_block (munge_mac_t md, const void *key, int keylen,
     }
     if ((e = gcry_md_open (&ctx, algo, GCRY_MD_FLAG_HMAC)) != 0) {
         log_msg (LOG_DEBUG, "gcry_md_open failed for MAC=%d HMAC: %s",
-            md, gcry_strerror (e));
+            (int) md, gcry_strerror (e));
         return -1;
     }
     if ((e = gcry_md_setkey (ctx, key, keylen)) != 0) {
         log_msg (LOG_DEBUG, "gcry_md_setkey failed for MAC=%d HMAC: %s",
-            md, gcry_strerror (e));
+            (int) md, gcry_strerror (e));
         gcry_md_close (ctx);
         return -1;
     }
